requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.4.0
drf-spectacular==0.27.2
